+++ lib/src/app.cpp
@@ -9,7 +9,7 @@ App::App(std::string tag)
 	  m_audio_streamer(std::make_unique<AudioStreamer>(*m_audio_device)) {
 	log::get_thread_id(); // set thread 0
 
-	for (int i = 0; i < static_cast<int>(m_gamepads.size()); ++i) { m_gamepads.at(static_cast<std::size_t>(i)).id = GamepadId{i}; }
+	for (int i = 0; i < static_cast<int>(m_gamepads.size()); ++i) { m_gamepads.at(static_cast<GamepadId>(i)).id = GamepadId{i}; }
 }
 
 void App::set_bootloader(Bootloader bootloader) {