+++ lib/docs/changelog.md
@@ -7,7 +7,7 @@
 - Added bave::Gamepad support, queried through bave::App::get_gamepad() and bave::App::get_gamepads().
   - Not yet supported on Android.
 - Added bave::KeyState and bave::App::get_key_state().
-- Added bave::EnumFlags.
+- Added bave::EnumFlags and bave::EnumArray.
 
 ### Changes